Moved snet and pathmgr to lib/api
