+++ go/lib/api/pathmgr/filter_map.go
