+++ go/lib/api/pathmgr/pathmgr.go
