+++ go/lib/api/pathmgr/pathmgr_test.go
