+++ go/lib/api/pathmgr/revtable.go
