+++ go/lib/api/pathmgr/syncpaths.go
