+++ go/lib/api/pathmgr/syncpaths_test.go
