+++ go/lib/api/snet/addr_test.go
