+++ go/lib/api/snet/integration_test.go
