+++ go/sig/base/pollhdlr.go
@@ -17,10 +17,10 @@ package base
 import (
 	log "github.com/inconshreveable/log15"
 
+	"github.com/netsec-ethz/scion/go/lib/api/snet"
 	"github.com/netsec-ethz/scion/go/lib/common"
 	"github.com/netsec-ethz/scion/go/lib/ctrl"
 	liblog "github.com/netsec-ethz/scion/go/lib/log"
-	"github.com/netsec-ethz/scion/go/lib/snet"
 	"github.com/netsec-ethz/scion/go/sig/disp"
 	"github.com/netsec-ethz/scion/go/sig/mgmt"
 	"github.com/netsec-ethz/scion/go/sig/sigcmn"
