+++ go/sig/egress/sesspath.go
@@ -21,7 +21,7 @@ import (
 
 	//log "github.com/inconshreveable/log15"
 
-	"github.com/netsec-ethz/scion/go/lib/pathmgr"
+	"github.com/netsec-ethz/scion/go/lib/api/pathmgr"
 	"github.com/netsec-ethz/scion/go/lib/sciond"
 )
