+++ go/sig/siginfo/sig.go
@@ -23,7 +23,7 @@ import (
 	//log "github.com/inconshreveable/log15"
 
 	"github.com/netsec-ethz/scion/go/lib/addr"
-	"github.com/netsec-ethz/scion/go/lib/snet"
+	"github.com/netsec-ethz/scion/go/lib/api/snet"
 )
 
 const (
